Fix static linking on iOS
ExternalInterface.cpp is used to register primitives for use by hxcpp's CFFI system. This is done using C++ static initializers. The ExternalInterface class, however, wasn't referenced by any other class, so the linker was free to strip it from builds, causing the primitives not to be registered, and games to crash when any of the unregistered functions were called.

Deal with this by calling the share_register_prims function from Share.mm.